// src/app/modules/core/direct-business/models/doiModel.ts
export class ProposerDetails {
    riskLocation: string;
    address: string;
    pinCode: string;
}

export class PartThreePolicyDetails {
    premiumStartDate: string;
    premiumEndDate: string;
    premium: string;
    claims: string;
}

export class BuildingWiseValue {
    descriptionOfBlock1: string;
    buildingIncludingPath: string;
    ma: string;
    ffOtherEquipment: string;
    ssp: string;
    propertyInsuredSeparately: string;
    total: string;
    age: string;
    height: string;
    construction: string;
}

export class TotalSumInsured {
    particulars: string;
    clausePerilCode: string;
    riskCode: string;
    rateCode: string;
    rate: string;
    sumInsured: string;
    premium: string;
    riskCode1: string;
    rateCode1: string;
    code?: string;
}